Add the `--compress` option
And make `--bucket` optional. This allows one to make a local backup
without the overhead of compressing it or uploading it if it's not
needed. Update cleanup to only delete temporary files. If `--bucket` is
set, that means all files. If it's not set but `--compress` is true, it
deletes the directory that was compressed and leaves the tarball. If
neither `--bucket` nor `--compress` is set, the directory is left.