Add `--cd`, improve connection options and vars
Add the `--cd` option, which tells Temback to switch to a directory
before starting the backup.

Improve the handling of the `--user`, `--pass`, and `--host` options and
their environment variable equivalents, `PGUSER`, `PGPASSWORD`, and
`PGHOST`. They no longer require values one way or the other, and
environment variables are no longer copied to connection parameters.